+++ client/components/Timeline/util.ts
@@ -1,3 +1,5 @@
 import React from 'react';
 
-export const TimelineContext = React.createContext({ accentColor: null });
+export const TimelineContext = React.createContext<{ accentColor: string | undefined }>({
+	accentColor: undefined,
+});
