+++ server/pub/queryMany.ts
@@ -2,8 +2,8 @@ import { QueryTypes, Op } from 'sequelize';
 import { QueryBuilder } from 'knex';
 
 import { knex, sequelize, Pub } from 'server/models';
-import { buildPubOptions, sanitizePub, SanitizedPubData } from 'server/utils/queryHelpers';
-import { InitialData, PubsQuery, PubGetOptions } from 'types';
+import { buildPubOptions, sanitizePub } from 'server/utils/queryHelpers';
+import { InitialData, PubsQuery, PubGetOptions, SanitizedPubData } from 'types';
 
 const defaultColumns = {
 	pubId: 'Pubs.id',
