+++ server/utils/queryHelpers/communityOverview.ts
@@ -1,10 +1,9 @@
 import { Collection, ScopeSummary } from 'server/models';
 import { getManyPubs } from 'server/pub/queryMany';
 import { getUserScopeVisits } from 'server/userScopeVisit/queries';
-import { Collection as CollectionType, InitialData } from 'types';
+import { Collection as CollectionType, InitialData, SanitizedPubData } from 'types';
 
 import sanitizeCollection from './collectionSanitize';
-import { SanitizedPubData } from './pubSanitize';
 
 type Options = {
 	loadPubs?: number;
